move code from ibis/pandas/api.py to ibis/pandas/__init__.py and refactor other files calling ibis.pandas.api
