+++ ibis/file/parquet.py
@@ -10,7 +10,7 @@ import ibis.expr.operations as ops
 import ibis.expr.schema as sch
 import ibis.expr.types as ir
 from ibis.file.client import FileClient
-from ibis.pandas.api import PandasDialect
+from ibis.pandas import PandasDialect
 from ibis.pandas.core import execute, execute_node
 
 dialect = PandasDialect
